moves files into new submodules
