+++ imagepypelines/core/io/CameraCapture.py
