+++ imagepypelines/core/io/Emailer.py
