+++ imagepypelines/core/io/ImageWriter.py
